fix api export so it can be imported by ts files
