+++ packages/server/test/integration/cypress_spec.js
@@ -24,7 +24,7 @@ const ciProvider = require(`../../lib/util/ci_provider`)
 const settings = require(`../../lib/util/settings`)
 const Windows = require(`../../lib/gui/windows`)
 const interactiveMode = require(`../../lib/modes/interactive`)
-const api = require(`../../lib/cloud/api`)
+const api = require(`../../lib/cloud/api`).default
 const cwd = require(`../../lib/cwd`)
 const user = require(`../../lib/cloud/user`)
 const cache = require(`../../lib/cache`)
