+++ packages/server/test/unit/cloud/api/api_spec.js
@@ -13,7 +13,7 @@ const {
   agent,
 } = require('@packages/network')
 const pkg = require('@packages/root')
-const api = require('../../../../lib/cloud/api')
+const api = require('../../../../lib/cloud/api').default
 const cache = require('../../../../lib/cache')
 const errors = require('../../../../lib/errors')
 const machineId = require('../../../../lib/cloud/machine_id')
@@ -237,7 +237,7 @@ describe('lib/cloud/api', () => {
 
       if (!prodApi) {
         prodApi = stealthyRequire(require.cache, () => {
-          return require('../../../../lib/cloud/api')
+          return require('../../../../lib/cloud/api').default
         }, () => {
           require('../../../../lib/cloud/encryption')
         }, module)
