+++ packages/server/test/unit/cloud/exceptions_spec.js
@@ -2,7 +2,7 @@ require('../../spec_helper')
 
 delete global.fs
 
-const api = require('../../../lib/cloud/api')
+const api = require('../../../lib/cloud/api').default
 const user = require('../../../lib/cloud/user')
 const exception = require('../../../lib/cloud/exception')
 const system = require('../../../lib/util/system')
